+++ client/src/features/auth/api/use-refresh.ts
@@ -2,7 +2,7 @@ import api from '@/lib/axios-instance';
 import { RefreshResponse } from '../auth.types';
 
 const refresh = async (): Promise<RefreshResponse> => {
-  const response = await api.post<RefreshResponse>('/api/auth/refresh');
+  const response = await api.post<RefreshResponse>('/api/auth/refresh', {});
   return response.data;
 };